[inductor][mps] Correctness fixes for the triton MPS backend
Make the triton MPS inductor path (config.mps_backend='triton', the
out-of-tree AppleGPU backend) compute correct results across the
test_torchinductor GPUTests suite:

- get_current_backend: report 'triton' for the mps device when
  mps_backend='triton', so upcast_compute_type applies the fp16->fp32
  reduction-accumulator upcasts (fixed 'Loop-carried variable ... type'
  compile errors in prod/argmax).
- to_dtype: an explicit narrowing float cast (.to(half)) must round through
  the low-precision type even when the compute type is upcast, otherwise the
  rounding is dropped (fp16 operand inside an fp32 cat diverged from eager).
- SPLIT_SCAN BackendFeature: decoupled-lookback split scan needs 64-bit
  device atomics, which Metal lacks; gate it off for mps so large scans fall
  back to a single cooperative scan instead of hanging the GPU.
- xfail_if_mps: only xfail on the native metal backend; the triton path
  generates Triton code and handles negative-index wrapping, kernel-count,
  etc. Split xfail_if_mps_unimplemented (missing eager MPS op -> always
  xfail) from xfail_if_mps (native-metal codegen quirk -> run on triton).

+++ torch/_inductor/fx_passes/joint_graph.py
@@ -1164,8 +1164,12 @@ def scatter_upon_const_tensor(
         # Create a mask for where to scatter
         mask = selector_expanded == indices_view
 
-        # Use torch.where to implement the scatter pointwise operation
-        return torch.where(mask, val, background_val)
+        # Use torch.where to implement the scatter pointwise operation.
+        # Both branches are python scalars, so torch.where would otherwise
+        # produce the default float dtype (f32) and silently widen a
+        # low-precision const tensor; round back to the original dtype to
+        # match eager scatter.value, which keeps the self tensor's dtype.
+        return torch.where(mask, val, background_val).to(dtype)
 
     # replace the scatter operation with pointwise equivalent
     # pyrefly: ignore [bad-argument-type]
